Phase 1: Setup Shared Frontend Types
